Update init_nvim script
extract appimage if it's in wsl.


Former-commit-id: 9238d0717c313f3050bcd3113a853fc30f159d53

+++ init_nvim.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
-
 sudo curl -fL https://github.com/neovim/neovim/releases/download/v0.3.0/nvim.appimage -o /opt/nvim.appimage
 sudo chmod +x /opt/nvim.appimage
-cd /opt/
-sudo rm -rf /opt/squashfs-root
-sudo /opt/nvim.appimage --appimage-extract
-sudo rm -rf /opt/neovim
-sudo mv squashfs-root neovim
-cd /usr/local/bin
-sudo find /opt/neovim/ -type d -exec chmod go+rx {} \;
-sudo ln -s /opt/neovim/AppRun nvim
 
+if grep -q Microsoft /proc/version; then
+  #wsl
+  cd /opt/
+  sudo rm -rf /opt/squashfs-root
+  sudo /opt/nvim.appimage --appimage-extract
+  sudo rm -rf /opt/neovim
+  sudo mv squashfs-root neovim
+  cd /usr/local/bin
+  sudo find /opt/neovim/ -type d -exec chmod go+rx {} \;
+  sudo ln -s /opt/neovim/AppRun nvim
+else
+  cd /usr/local/bin
+  sudo ln -s /opt/nvim.appimage nvim
+fi
 mv ~/.config/nvim/init.vim ~/.config/nvim/init.vim.`date +%F_%R`
 
 curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
